+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+- Add new request: list items linked to a collection.
+- Add new request: Create a job from its workflow.
+- Add new request: Update a collection's item.
+- Add new request: Update a job's cell.
+- Add new request: Delete a collection's item.
+- Fetch the details for the given user.
+
 ## [0.4.0] - 2023-09-12
 
 - Require `AddTagRequest`